+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [7.7.0] - 2024-11-12
 ### Changed
 - Updated the hms namespaces for metrics for both readwrite and readonly.
+### Added
+- Added following variables:
+  - `hms_ecs_metrics_readwrite_namespace`
+  - `hms_ecs_metrics_readonly_namespace`
+  - `hms_k8s_metrics_readwrite_namespace`
+  - `hms_k8s_metrics_readonly_namespace`
+  - `hms_ro_tolerations`
 
 ## [7.6.1] - 2024-10-30
 ### Changed